refactor
move csharpproject commands and events in the right project
ensure all commands and events are tested
ensure api contract tests check csharp project too
