+++ src/Microsoft.DotNet.Interactive.ApiCompatibility.Tests/ApiCompatibilityTests.Interactive_api_is_not_changed.approved.txt
@@ -310,13 +310,6 @@ Microsoft.DotNet.Interactive
   public class PasswordString
     .ctor(System.String clearTextPassword)
     public System.String GetClearTextPassword()
-  public class Project
-    .ctor(System.Collections.Generic.IReadOnlyCollection<ProjectFile> files)
-    public System.Collections.Generic.IReadOnlyCollection<ProjectFile> Files { get;}
-  public class ProjectFile
-    .ctor(System.String path, System.String content)
-    public System.String Content { get;}
-    public System.String Path { get;}
   public class ResolvedPackageReference : PackageReference
     .ctor(System.String packageName, System.String packageVersion, System.Collections.Generic.IReadOnlyList<System.String> assemblyPaths, System.String packageRoot = null, System.Collections.Generic.IReadOnlyList<System.String> probingPaths = null)
     public System.Collections.Generic.IReadOnlyList<System.String> AssemblyPaths { get;}
@@ -349,9 +342,6 @@ Microsoft.DotNet.Interactive.Commands
     .ctor(System.String workingDirectory)
     public System.String WorkingDirectory { get;}
     public System.Threading.Tasks.Task InvokeAsync(Microsoft.DotNet.Interactive.KernelInvocationContext context)
-  public class CompileProject : KernelCommand
-    .ctor(System.String code = null)
-    public System.String Code { get;}
   public class DisplayError : KernelCommand
     .ctor(System.String message)
     public System.String Message { get;}
@@ -374,13 +364,6 @@ Microsoft.DotNet.Interactive.Commands
     public Microsoft.DotNet.Interactive.LinePosition LinePosition { get;}
     protected System.Void set_Code(System.String value)
     protected System.Void set_LinePosition(Microsoft.DotNet.Interactive.LinePosition value)
-  public class OpenDocument : KernelCommand
-    .ctor(System.String path, System.String regionName = null)
-    public System.String Path { get;}
-    public System.String RegionName { get;}
-  public class OpenProject : KernelCommand
-    .ctor(Microsoft.DotNet.Interactive.Project project)
-    public Microsoft.DotNet.Interactive.Project Project { get;}
   public class Quit : KernelCommand
     .ctor()
     public System.Threading.Tasks.Task InvokeAsync(Microsoft.DotNet.Interactive.KernelInvocationContext context)
@@ -487,9 +470,6 @@ Microsoft.DotNet.Interactive.Connection
     public Microsoft.DotNet.Interactive.ValueSharing.IKernelValueDeclarer GetValueDeclarer(System.Object value)
     public System.Threading.Tasks.Task StartAsync()
 Microsoft.DotNet.Interactive.Events
-  public class AssemblyProduced : KernelEvent
-    .ctor(Microsoft.DotNet.Interactive.Commands.CompileProject command, Microsoft.DotNet.Interactive.Base64EncodedAssembly assembly)
-    public Microsoft.DotNet.Interactive.Base64EncodedAssembly Assembly { get;}
   public class CodeSubmissionReceived : KernelEvent
     .ctor(Microsoft.DotNet.Interactive.Commands.SubmitCode command)
     public System.String Code { get;}
@@ -543,11 +523,6 @@ Microsoft.DotNet.Interactive.Events
     public System.Object Value { get;}
     public System.String ValueId { get;}
     public System.String ToString()
-  public class DocumentOpened : KernelEvent
-    .ctor(Microsoft.DotNet.Interactive.Commands.OpenDocument command, System.String path, System.String regionName, System.String content)
-    public System.String Content { get;}
-    public System.String Path { get;}
-    public System.String RegionName { get;}
   public class ErrorProduced : DisplayEvent
     .ctor(System.String message, Microsoft.DotNet.Interactive.Commands.KernelCommand command, System.Collections.Generic.IReadOnlyCollection<Microsoft.DotNet.Interactive.FormattedValue> formattedValues = null)
     public System.String Message { get;}
@@ -703,6 +678,7 @@ Microsoft.DotNet.Interactive.Server
     public static IKernelCommandEnvelope Create(Microsoft.DotNet.Interactive.Commands.KernelCommand command)
     public static IKernelCommandEnvelope Deserialize(System.String json)
     public static IKernelCommandEnvelope Deserialize(System.Text.Json.JsonElement json)
+    public static System.Void RegisterCommand<T>()
     public static System.Void RegisterDefaults()
     public static System.String Serialize(Microsoft.DotNet.Interactive.Commands.KernelCommand command)
     public static System.String Serialize(IKernelCommandEnvelope envelope)
