+++ src/Microsoft.DotNet.Interactive.ApiCompatibility.Tests/Microsoft.DotNet.Interactive.ApiCompatibility.Tests.csproj
@@ -44,6 +44,7 @@
 
 
   <ItemGroup>
+    <ProjectReference Include="..\Microsoft.DotNet.Interactive.CSharpProject\Microsoft.DotNet.Interactive.CSharpProject.csproj" />
     <ProjectReference Include="..\Microsoft.DotNet.Interactive.CSharp\Microsoft.DotNet.Interactive.CSharp.csproj" />
     <ProjectReference Include="..\Microsoft.DotNet.Interactive.Documents\Microsoft.DotNet.Interactive.Documents.csproj" />
     <ProjectReference Include="..\Microsoft.DotNet.Interactive.Formatting\Microsoft.DotNet.Interactive.Formatting.csproj" />
