+++ src/Microsoft.DotNet.Interactive.CSharpProject/Commands/CompileProject.cs
@@ -1,0 +1,16 @@
+﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using Microsoft.DotNet.Interactive.Commands;
+
+namespace Microsoft.DotNet.Interactive.CSharpProject.Commands;
+
+public class CompileProject : KernelCommand
+{
+    public string Code { get; }
+
+    public CompileProject(string code = null)
+    {
+        Code = code;
+    }
+}
