+++ src/Microsoft.DotNet.Interactive.CSharpProject/Commands/OpenProject.cs
@@ -1,0 +1,17 @@
+﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System;
+using Microsoft.DotNet.Interactive.Commands;
+
+namespace Microsoft.DotNet.Interactive.CSharpProject.Commands;
+
+public class OpenProject : KernelCommand
+{
+    public Project Project { get; }
+
+    public OpenProject(Project project)
+    {
+        Project = project ?? throw new ArgumentNullException(nameof(project));
+    }
+}
