+++ src/Microsoft.DotNet.Interactive.Tests/Server/SerializationTests.Command_contract_has_not_been_broken.approved.GetInput.json
@@ -1,0 +1,1 @@
+{"token":"the-token","id":"command-id","commandType":"GetInput","command":{"prompt":"provide answer","isPassword":true,"targetKernelName":"vscode"},"originUri":"","destinationUri":""}
