+++ src/Microsoft.DotNet.Interactive.Tests/Server/SerializationTests.Event_contract_has_not_been_broken.approved.InputProduced.json
@@ -1,0 +1,1 @@
+{"event":{"value":"user input"},"eventType":"InputProduced","command":{"token":"the-token","id":"command-id","commandType":"GetInput","command":{"prompt":"","isPassword":false,"targetKernelName":"vscode"},"originUri":"","destinationUri":""}}
