--- src/Microsoft.DotNet.Interactive/ProjectFile.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Microsoft.DotNet.Interactive
-{
-    public class ProjectFile
-    {
-        public string Path { get; }
-        public string Content { get; }
-
-        public ProjectFile(string path, string content)
-        {
-            Path = path;
-            Content = content;
-        }
-    }
-}
